refactor: moved the code into folders for better strcuture
