+++ src/bmiemg/data/convert/__init__.py
@@ -1,5 +1,5 @@
 from .session_loader import session_load
-from .ChannelSplitter import ChannelSplitter
+from .spliter import ChannelSplitter
 
 __all__ = [
     "session_load",
