+++ src/bmiemg/data/convert/spliter/ChannelMap.py
