+++ src/bmiemg/data/convert/spliter/__init__.py
@@ -1,0 +1,11 @@
+from .ChannelSplitter import ChannelSplitter
+from .ChannelMap import (
+    ChannelMap,
+    BIOTECH_MAP
+)
+
+__all__ = [
+    "ChannelMap",
+    "BIOTECH_MAP",
+    "ChannelSplitter"
+]
