+++ src/bmiemg/data/convert/streams/SignalStream.py
@@ -46,7 +46,7 @@ class SignalStream(Stream):
                 "To raw only works when the signal is mono (only one modality)"
             )
 
-        data = self.time_series.T
+        data = (self.time_series * 1e-6).T
 
         mne_info = mne.create_info(
             ch_names = list(self.channel_names),
