+++ src/bmiemg/data/convert/streams/Stream.py
