+++ src/bmiemg/data/convert/streams/__init__.py
@@ -1,0 +1,7 @@
+from .SignalStream import SignalStream
+from .MarkerStream import MarkerStream
+
+__all__ = [
+    "SignalStream",
+    "MarkerStream"
+]
